find_adversarial_example: enable searching for worst adversarial example (#6)
find_adversarial_example previously searched for the closest adversarial example. We now enable searching for the worst adversarial example via the optional parameter adversarial_example_objective in find_adversarial_example.

When adversarial_example_objective = MIPVerify.worst, the optimizer will attempt to find the adversarial example where the difference between the maximum target activation [1] and the maximum off-target activation [2] is maximized. (Note that when adversarial_example_objective=MIPVerify.worst, the norm_order parameter specified is ignored.) The objective value is positive if the maximum target activation exceeds the maximum off-target activation.

The same parameter is available in batch_find_untargeted_attack, which is how we expect users to call the function.