Bugfix: Guarantee that `find_adversarial_example` output is in target category.
This change is inspired by #34.

Prior to the PR, it was possible for the `find_adversarial_example` to find an "adversarial example" that was not, in fact, classified in the `target_selection`. This is contrary to what the documentation promised.

This PR fixes that and also makes some incidental formatting changes.